all test pass

+++ src/main/java/com/swagger/swaggerapi/todo/MessageRepo.java
@@ -12,6 +12,6 @@ import org.springframework.data.repository.CrudRepository;
  *
  * @author Jun.Meng
  */
-public interface MessageRepo extends CrudRepository<Message, Long>  {
+public interface MessageRepo extends CrudRepository<Message, Integer>  {
     
 }
